requirements.txt
Flask==1.1.2
pylint==2.5.3
python-dotenv==0.13.0
psycopg2